refactor(input): remove unnecessary declarations in the header
